move checks to own dir
